update capture.sh to set 8 times speed

+++ scripts/capture.sh
@@ -5,7 +5,7 @@
 #   $ sudo apt-get install -y ffmpeg
 
 # default value
-FPS="3"          # capture frame per sec
+FPS="4"          # capture frame per sec
 CAPTURE_NAME="capture.mp4"
 MODE="start"
 
@@ -54,10 +54,20 @@ elif [ $MODE == "stop" ];then
     # kill process
     PROCESS_ID=`ps -e -o pid,cmd | grep "recordmydesktop" | grep -v grep | awk '{print $1}'`
     kill $PROCESS_ID
-    sleep 30
+    # wait to kill process
+    LOOP_TIMES=3600
+    for i in `seq ${LOOP_TIMES}`
+    do
+	NUM=`ps -e -o pid,cmd | grep "recordmydesktop" | grep -v grep | wc -l`
+	if [ $NUM -eq 0 ]; then
+	    break;
+	fi
+	sleep 1
+    done    
     # save video
-    ffmpeg -i $TMP_FNAME_OGV -vf scale=1024:-1 $TMP_FNAME_MP4
-    ffmpeg -i $TMP_FNAME_MP4 -vf setpts=PTS/3.3 -af atempo=2.0 $TMP2_FNAME_MP4
+    #ffmpeg -i $TMP_FNAME_OGV -vf scale=1024:-1 $TMP_FNAME_MP4
+    #ffmpeg -i $TMP_FNAME_MP4 -vf setpts=PTS/3.3 -af atempo=2.0 $TMP2_FNAME_MP4
+    ffmpeg -i $TMP_FNAME_OGV -vf setpts=PTS/8 -af atempo=2.0 $TMP2_FNAME_MP4
     mv $TMP2_FNAME_MP4 ${CAPTURE_NAME}
     rm -f $TMP_FNAME_OGV
     rm -f $TMP_FNAME_MP4
